ETL Logging
This enables ETL logging to Trace for debug builds. Also, fixes a bug where
etl_profile.h wasn't actually getting included by ETL (and thus, was having
no actual effect on ETL's behavior).

Basically, ETL does

	#include "etl_profile.h"

So the `config` path needs to be on the list of top level include directories.

I'm not great at CMake so I'm not sure if this is the right way to do that.

Closes #443