Fix i2s_audio to sign extend correctly
The i2s audio code as written was not correctly sign-extending
twos-complement values, because it was padding with zeroes between the
MSB and the rest of the value regardless of the MSB value. The correct
way to sign extend is to pad with 1s for a negative number and 0s for
a positive (i.e. replicate the sign bit)

However, thankfully, the code was inadvertently sign-extending anyway
because the data pins for `set` were not configured at all, so the set
command was essentially acting as a NOP, preserving the MSB during the
padding.

This commit fixes the code to explicitly sign-extend correctly.