Work around DAC chip auto-mute
The DAC chip used in the picotracker, pcm5102a, has an auto-mute
behavior when it detects a continuous string of zeroes. It is described
in section 9.3.2.3 of the datasheet.

For whatever reason, when this kicks in, it causes an audible click
sound, which is particularly apparent when auditioning notes while
sequencing, or during quiet sections of a song that may go to silence
for a row or two.

It's possible to prevent this by inserting a 1 in the LSB of the 32-bit
value we send in the i2s datastream. This should be well below the
noise floor, so it should not cause any harm to the actual audio signal.