i2s: sign extend correctly; insert 1-bit padding to prevent DAC automute (#449)
* Fix i2s_audio to sign extend correctly

The i2s audio code as written was not correctly sign-extending
twos-complement values, because it was padding with zeroes between the
MSB and the rest of the value regardless of the MSB value. The correct
way to sign extend is to pad with 1s for a negative number and 0s for
a positive (i.e. replicate the sign bit)

However, thankfully, the code was inadvertently sign-extending anyway
because the data pins for `set` were not configured at all, so the set
command was essentially acting as a NOP, preserving the MSB during the
padding.

This commit fixes the code to explicitly sign-extend correctly.

* Work around DAC chip auto-mute

The DAC chip used in the picotracker, pcm5102a, has an auto-mute
behavior when it detects a continuous string of zeroes. It is described
in section 9.3.2.3 of the datasheet.

For whatever reason, when this kicks in, it causes an audible click
sound, which is particularly apparent when auditioning notes while
sequencing, or during quiet sections of a song that may go to silence
for a row or two.

It's possible to prevent this by inserting a 1 in the LSB of the 32-bit
value we send in the i2s datastream. This should be well below the
noise floor, so it should not cause any harm to the actual audio signal.