Preserve column when navigating up/down in grid-layout FieldViews
Multi-column views such as OPAL had the frustrating characteristic that
pressing up or down would reset the column to the leftmost column in
the grid. This code makes the focus shifting logic column-aware when
navigating up or down, snapping to the column with the most-similar X
value in the next selected row.

This produces perfect results when navigating columns where the X values
match, and decent results when moving between grids with different
numbers of columns on a single FieldView.